# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Task Board API",
        "version": "0.0.1",
        "description": "\u30c1\u30fc\u30e0\u5171\u6709\u30bf\u30b9\u30af\u30dc\u30fc\u30c9\u306e API\u3002\u8a8d\u8a3c\u306f Sanctum \u306e API \u30c8\u30fc\u30af\u30f3\uff08`Authorization: Bearer <token>`\uff09\u3002"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/projects/{project}": {
            "get": {
                "operationId": "projects.show",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "projects.update",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "projects.destroy",
                "summary": "\u30e1\u30f3\u30d0\u30fc\u3068\u30bf\u30b9\u30af\u306f\u5916\u90e8\u30ad\u30fc\u306e ON DELETE CASCADE \u3067\u4e00\u7dd2\u306b\u6d88\u3048\u308b\u3002",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/projects/{project}/members/{member}": {
            "patch": {
                "operationId": "projects.members.update",
                "tags": [
                    "ProjectMember"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member user id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectMemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProjectMemberResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectMemberResource"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "projects.members.destroy",
                "summary": "owner \u304c\u5916\u3059\u5834\u5408\u3068\u3001\u672c\u4eba\u304c\u629c\u3051\u308b\u5834\u5408\u306e\u4e21\u65b9\u3092\u3053\u3053\u3067\u53d7\u3051\u308b\u3002",
                "tags": [
                    "ProjectMember"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member user id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/projects/{project}/tasks/{task}": {
            "get": {
                "operationId": "projects.tasks.show",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "projects.tasks.update",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`TaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "projects.tasks.destroy",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/auth/two-factor": {
            "post": {
                "operationId": "twoFactorAuthentication.store",
                "summary": "\u767b\u9332\u3092\u958b\u59cb\u3059\u308b\u3002\u30b7\u30fc\u30af\u30ec\u30c3\u30c8\u3092\u767a\u884c\u3059\u308b\u304c\u3001\u3053\u306e\u6642\u70b9\u3067\u306f MFA \u306f\u307e\u3060\u7121\u52b9\u3002",
                "tags": [
                    "TwoFactorAuthentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordConfirmationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "secret": {
                                            "type": "string",
                                            "description": "QR \u753b\u50cf\u306f\u4f5c\u3089\u305a otpauth: \u306e URI \u3060\u3051\u3092\u8fd4\u3059\u3002\u753b\u50cf\u5316\u306f\u30d5\u30ed\u30f3\u30c8\u306e\u8cac\u52d9\n\uff08\u8868\u793a\u306e\u90fd\u5408\u3067\u89e3\u50cf\u5ea6\u3084\u914d\u8272\u304c\u5909\u308f\u308b\u4e00\u65b9\u3001URI \u306f\u5909\u308f\u3089\u306a\u3044\uff09\u3002\n\u624b\u5165\u529b\u3057\u305f\u3044\u30e6\u30fc\u30b6\u30fc\u5411\u3051\u306b secret \u3082\u8fd4\u3059\u3002\u3069\u3061\u3089\u3082\u540c\u3058\u9375\u3092\u904b\u3076\u3002"
                                        },
                                        "otpauth_uri": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "secret",
                                        "otpauth_uri"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "examples": [
                                                "Two factor authentication is already enabled."
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "twoFactorAuthentication.destroy",
                "summary": "MFA \u3092\u89e3\u9664\u3059\u308b\u3002\u78ba\u8a8d\u5f85\u3061\u306e\u767b\u9332\u3092\u53d6\u308a\u6d88\u3059\u7528\u9014\u3082\u517c\u306d\u308b\u3002",
                "tags": [
                    "TwoFactorAuthentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "current_password \u306f\u8a8d\u8a3c\u4e2d\u306e\u30e6\u30fc\u30b6\u30fc\u306e\u30cf\u30c3\u30b7\u30e5\u3068\u7167\u5408\u3059\u308b\uff08Hash::check \u76f8\u5f53\uff09\u3002\n\u30ac\u30fc\u30c9\u3092\u660e\u793a\u3059\u308b\u306e\u306f\u3001\u65e2\u5b9a\u30ac\u30fc\u30c9\u306b\u4f9d\u5b58\u3055\u305b\u306a\u3044\u305f\u3081\u3002auth:sanctum \u306f\u8a8d\u8a3c\u304c\n\u901a\u3063\u305f\u3068\u304d\u306b shouldUse('sanctum') \u3092\u547c\u3076\u304c\u3001\u305d\u308c\u306b\u983c\u308b\u3068\u8a8d\u8a3c\u306e\u8a2d\u5b9a\u3092\n\u5909\u3048\u305f\u3068\u304d\u306b\u7167\u5408\u5148\u304c\u9759\u304b\u306b\u305a\u308c\u308b\u3002"
                                    }
                                },
                                "required": [
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ProjectMemberResource": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/ProjectRole"
                    },
                    "joined_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "user_id",
                    "name",
                    "email",
                    "role",
                    "joined_at"
                ],
                "title": "ProjectMemberResource"
            },
            "ProjectResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "role": {
                        "$ref": "#/components/schemas/ProjectRole"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description",
                    "role",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProjectResource"
            },
            "ProjectRole": {
                "type": "string",
                "description": "| |\n|---|\n| `owner` <br/> \u30d7\u30ed\u30b8\u30a7\u30af\u30c8\u306e\u7ba1\u7406\u8005\u3002\u66f4\u65b0\u30fb\u524a\u9664\u30fb\u30e1\u30f3\u30d0\u30fc\u7ba1\u7406\u304c\u3067\u304d\u308b |\n| `member` <br/> \u53c2\u52a0\u8005\u3002\u30bf\u30b9\u30af\u306f\u64cd\u4f5c\u3067\u304d\u308b\u304c\u3001\u30d7\u30ed\u30b8\u30a7\u30af\u30c8\u81ea\u4f53\u306f\u5909\u66f4\u3067\u304d\u306a\u3044 |",
                "enum": [
                    "owner",
                    "member"
                ],
                "title": "ProjectRole"
            },
...
            "TaskResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "$ref": "#/components/schemas/TaskStatus"
                    },
                    "due_date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "date \u30ad\u30e3\u30b9\u30c8\u306e\u5024\u3092\u305d\u306e\u307e\u307e\u51fa\u3059\u3068 \"2026-10-05T00:00:00.000000Z\" \u306b\u306a\u308a\u3001\n\u53d7\u3051\u53d6\u3063\u305f\u5074\u306e\u30bf\u30a4\u30e0\u30be\u30fc\u30f3\u3067\u524d\u65e5\u306b\u305a\u308c\u3046\u308b\u3002\u65e5\u4ed8\u3060\u3051\u306e\u5024\u3068\u3057\u3066\u51fa\u3059\u3002"
                    },
                    "assignee": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "\u30ab\u30fc\u30c9\u306b\u62c5\u5f53\u8005\u306e\u540d\u524d\u3092\u51fa\u3059\u305f\u3081\u57cb\u3081\u8fbc\u3080\uff08GitHub \u306e Issue \u3082 assignee \u3092\u57cb\u3081\u8fbc\u3080\uff09\u3002\n\u30e1\u30fc\u30eb\u30a2\u30c9\u30ec\u30b9\u306f\u30e1\u30f3\u30d0\u30fc\u4e00\u89a7\u306b\u3042\u308b\u306e\u3067\u3001\u3053\u3053\u306b\u306f\u51fa\u3055\u306a\u3044\u3002",
                        "properties": {
                            "user_id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "user_id",
                            "name"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "description",
                    "status",
                    "due_date",
                    "assignee",
                    "created_at",
                    "updated_at"
                ],
                "title": "TaskResource"
            },
            "TaskStatus": {
                "type": "string",
                "enum": [
                    "todo",
                    "in_progress",
                    "done"
                ],
                "title": "TaskStatus"
            },
...
            "UpdateProjectMemberRequest": {
                "type": "object",
                "properties": {
                    "role": {
                        "$ref": "#/components/schemas/ProjectRole"
                    }
                },
                "required": [
                    "role"
                ],
                "title": "UpdateProjectMemberRequest"
            },
            "UpdateProjectRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 5000
                    }
                },
                "title": "UpdateProjectRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}